+++ Source/Common/FrontierDevelopments-Shields.csproj
@@ -83,6 +83,7 @@
     <Compile Include="Shields\Mod.cs" />
     <Compile Include="Shields\IShield.cs" />
     <Compile Include="Shields\Settings.cs" />
+    <Compile Include="Shields\ShieldDamage.cs" />
     <Compile Include="Shields\ShieldManager.cs" />
     <Compile Include="Shields\IShieldResists.cs" />
     <Compile Include="Shields\ShieldSetting.cs" />
